wise_evaluation/requirements.txt
# Python Dependencies
requests>=2.28.0
openai>=1.0.0
python-dotenv>=0.19.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
Pillow>=9.0.0
numpy>=1.24.0

# WISE Submodule Setup
# Official WISE benchmark is included as git submodule
# To initialize on new machine: git submodule update --init --recursive
# To update WISE: cd WISE && git pull origin main && cd .. && git add WISE
# WISE provides 1000 prompts across 25 subdomains for world knowledge evaluation 